Создала папки Categories, Recipes и Modal и переместила в них соответствующие файлы
