+++ src/components/Categories/AllRecipes.jsx
@@ -1,28 +1,27 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { Link, Route } from 'react-router-dom';
-import Recipes from './Recipes';
+import Recipes from '../Recipes/Recipes';
 
-function Beverages(props) {
+function AllRecipes(recipe) {
   const recipes = useSelector((state) => state.recipes.items);
-  const newRecipes = recipes.filter((item) => item.category === 'напитки');
 
   return (
     <div>
       <div>
-        <Route path="/beverages/:id?/:title?">
+        <Route path="/all-recipes/:id?/:title?">
           <Recipes />
         </Route>
       </div>
-      {newRecipes.map((item) => {
+      {recipes.map((item) => {
         return (
           <div
             style={{ width: '58%' }}
             className="recipes justify-content-center m-auto"
           >
             <div className="recipes-block">
-              <Route exact path="/beverages/:id?">
-                <Link to={`/beverages/${item.id}/${item.title}`}>
+              <Route exact path="/all-recipes/:id?">
+                <Link to={`/all-recipes/${item.id}/${item.title}`}>
                   <div className="d-inline-flex">
                     <div className="img mb-3">
                       <img width={286} height={180} src={item.url} alt="img" />
@@ -57,4 +56,4 @@ function Beverages(props) {
   );
 }
 
-export default Beverages;
+export default AllRecipes;
