+++ src/components/Main.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import Search from './Search';
 import Slider from './Slider';
-import './modal.css';
-import MainRecipes from './MainRecipes';
+import MainRecipes from './Recipes/MainRecipes';
 
 function Main(props) {
   return (
